Create route super class
